[SETTING] enum 분리 및 엔티티 수정
