+++ src/main/java/com/lunchchat/domain/chat_participant/entity/ChatParticipant.java
@@ -1,7 +1,7 @@
-package com.lunchchat.domain.chat_participant;
+package com.lunchchat.domain.chat_participant.entity;
 
-import com.lunchchat.domain.chat_room.ChatRoom;
-import com.lunchchat.domain.users.User;
+import com.lunchchat.domain.chat_room.entity.ChatRoom;
+import com.lunchchat.domain.users.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -9,9 +9,7 @@ import java.io.Serializable;
 
 @Entity
 @Getter
-@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
 @IdClass(ChatParticipantId.class)
 public class ChatParticipant implements Serializable {
