+++ src/main/java/com/lunchchat/domain/chat_room/entity/ChatRoom.java
@@ -1,19 +1,18 @@
-package com.lunchchat.domain.college;
+package com.lunchchat.domain.chat_room.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.time.LocalDateTime;
+
 @Entity
 @Getter
-@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class College {
+public class ChatRoom {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, unique = true)
-    private String name;
+    private LocalDateTime createdAt;
 }
