+++ src/main/java/com/lunchchat/domain/interests/entity/InterestName.java
@@ -1,0 +1,12 @@
+package com.lunchchat.domain.interests.entity;
+
+public enum InterestName {
+    EXCHANGE_STUDENT, // 교환학생
+    EMPLOYMENT_CAREER, // 취업/진로
+    EXAM_PREPARATION, // 고시준비
+    STARTUP, // 창업
+    FOREIGN_LANGUAGE_STUDY, // 외국어 공부
+    HOBBY_LEISURE, // 취미/여가
+    SCHOOL_LIFE, // 학교생활
+    ETC // 기타
+}
