+++ src/main/java/com/lunchchat/domain/interests/entity/Interests.java
@@ -1,0 +1,20 @@
+package com.lunchchat.domain.interests.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Interests {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false, unique = true)
+    private InterestName name;
+
+    
+}
