+++ src/main/java/com/lunchchat/domain/match/entity/MatchStatus.java
@@ -1,0 +1,7 @@
+package com.lunchchat.domain.match.entity;
+
+public enum MatchStatus {
+    REQUESTED,
+    ACCEPTED,
+    REJECTED
+}
