+++ src/main/java/com/lunchchat/domain/project_info/entity/ProjectInfo.java
@@ -1,22 +1,19 @@
-package com.lunchchat.domain.faq;
+package com.lunchchat.domain.project_info.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
 @Getter
-@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Faq {
+public class ProjectInfo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @Column(columnDefinition = "TEXT")
-    private String question;
+    private String title;
 
     @Column(columnDefinition = "TEXT")
-    private String answer;
+    private String content;
 }
