+++ src/main/java/com/lunchchat/domain/time_table/entity/DayOfWeek.java
@@ -1,0 +1,5 @@
+package com.lunchchat.domain.time_table.entity;
+
+public enum DayOfWeek {
+    MON, TUE, WED, THU, FRI, SAT, SUN
+}
