+++ src/main/java/com/lunchchat/domain/user_interests/entity/UserInterests.java
@@ -1,7 +1,7 @@
-package com.lunchchat.domain.user_interests;
+package com.lunchchat.domain.user_interests.entity;
 
-import com.lunchchat.domain.interests.Interests;
-import com.lunchchat.domain.users.User;
+import com.lunchchat.domain.interests.entity.Interests;
+import com.lunchchat.domain.users.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -9,9 +9,7 @@ import java.io.Serializable;
 
 @Entity
 @Getter
-@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
 @IdClass(UserInterestsId.class)
 public class UserInterests implements Serializable {
