+++ src/main/java/com/lunchchat/domain/user_keywords/entity/KeywordType.java
@@ -1,0 +1,7 @@
+package com.lunchchat.domain.user_keywords.entity;
+
+public enum KeywordType {
+    EXPRESS, // express
+    GOAL, // goal
+    INTEREST // interest
+}
